force reseed after 10 minutes
Due to the removal of the blocking pool in the kernel, it is quite
unlikely that an rngd is ever triggered because the kernel thinks it has
sufficient entropy (the ChaCha20 DRNG is reseeded every 5 minutes with
256 bits of entropy - an rngd is only triggered if the kernel is unable
to collect at least 256 bits of entropy within 5 minutes).

Thus, the jitterentropy-rngd will unconditionally inject 256 bits of
entropy every 10 minutes.

Signed-off-by: Stephan Mueller <smueller@chronox.de>