prevent creating timer thread if high-res timer is detected
When enabling JENT_CONF_ENABLE_INTERNAL_TIMER during compile time, the
additional code should be completely dead during run time if a high-res
hardware timer is detected.

Unfortunately that was not the case as the timer thread was always
created even though not needed. The thread is now only created if no
high-res hardware timer is detected.

Reported-by: Lonnie Abelbeck
Signed-off-by: Stephan Mueller <smueller@chronox.de>