SP800-90B compliance support
The existing /dev/random implementation uses a SHA-1 operation whose
output is folded in half when generating data from the input_pool. This
implies that according to SP800-90B section 3.1.5.1, the n_out is 80
bits and that this operation can only deliver 80 bits at a maximum in
one invocation.

To support the case that the Jitter RNG provides SP800-90B compliant
noise data and that the output of /dev/random can be claimed to have
full entropy, the RNGD injects the data into the kernel with the
following strategy:

- If the LRNG is present, take no special precautions as this operates
  SP800-90B compliant.

- If the kernel is >= 4.17, chunk the data into 80 bit blocks, inject
these blocks into the kernel and after each injection use the
RNDRESEEDCRNG IOCTL to force the transport of the 80 bits of entropy
into the ChaCha20. This apporach implies that the ChaCha20 DRNG is
reseeded several times with 80 bits up to 256 bits of entropy and thus
can be claimed to have full entropy.

- If the kernel is < 4.17, the IOCTL is not available. Thus the data is
  inserted completely into the kernel. Yet, if the Jitter RNG shall
  operate SP800-90B-compliant with the -s command line flag, only 80
  bits of entropy are injected at
  all. The idea now is that either by adjusting the /proc values to
  cause /dev/urandom to always seed from the input_pool or by using
  /dev/random, after reading 80 bits from the kernel, the RNGD is
  triggered again. This way, the caller invocation ensures the chunking
  of the entropy data into 80 bits. This behavior implies, however, that
  the RNGD is always woken up when even one bit of entropy is missing
  from the pool. This implies that
  /proc/sys/kernel/random/write_wakeup_threshold needs to be set to
  a value around 4095.

In addition, this patch set fixes a bug where the invocation of the
RNDRESEEDCRNG causes the RNGD to drain the kernel entropy pool leading
to an endless loop.

Signed-off-by: Stephan Mueller <smueller@chronox.de>