Inject sufficient entropy into input_pool
When using the reseed IOCTL, it is invoked multiple times. Each time it
is able to drain up to 256 bits from the entropy estimator. Thus, the
entropy estimator must be increased by more than the number of IOCTLs
times 256 bits to ensure that the RNGD always ensures the estimator
rises. Otherwise an endless loop is triggered.

In addition, the invocation of the reseed IOCTL is only applied during
startup and when requested with the -s command line flag.

Signed-off-by: Stephan Mueller <smueller@chronox.de>