JENT: add replacement if cache sysconf flags not available
If the sysconf calls for getting the cache size are not available, read
the data out of SysFS. The SysFS files for the CPU0 (which should exist
in all cases) are accessed and read. The read operation attempts to read
up to 3 cache entries, searching for data and unified caches. This
covers the same functionality as the sysconf calls.

If errors occur, the read cache sizes are zero which is appropriately
handled by the Jitter RNG code base.

Reported-by: stoickish
Signed-off-by: Stephan Mueller <smueller@chronox.de>