For SP800-90B ensure stable H_submitter
The H_submitter direclty depends on the selected oversampling rate.
The Jitter RNG API of jent_entropy_read_safe, however automatically
changes the OSR if there was a health failure. This implies that the
H_submitter changes.

When enabling the SP800-90B compliance mode, use jent_entropy_read
which fails directly on the health failure guaranteeing a stable
H_submitter.

Signed-off-by: Stephan Mueller <smueller@chronox.de>